Bringing Technology to Kids -The Bento Box
This is the code for Processing and it is complicated since it displays the game interface. In the Bento Box game, each player has 3 tries before the game is "over" when an image of a "Sad Sushi" appears. On the contrast, through every try until the 3rd one, there is an image of a "Happy Sushi" which encourages kids to keep trying.